chore: relocate Input
Input component relocated into its own directory
