+++ src/components/input/types.ts
@@ -1,0 +1,6 @@
+export interface InputProps
+  extends React.InputHTMLAttributes<HTMLInputElement> {
+  label?: string;
+  handleIncreaseValue?: () => void;
+  handleDecreaseValue?: () => void;
+}
